fix(mobile): resolve Android i18n sync and iOS SwiftLint CI failures
Regenerate Localizable.xcstrings from mobile locale sources and address
SwiftLint errors in quiz, grade calculator, and annotation model code.
